requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.7
django-ckeditor==6.1.0
django-fontawesome-5==1.0.18
django-js-asset==1.2.2
django-osm-field==0.4.1
django-tawkto==0.3
gunicorn==20.1.0
Pillow==8.3.2
psycopg==3.0b1
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
